Add the elf note build-id to the vmlinuz image for ppc64le
This is failing kernel.spec when invoking:

     + /usr/lib/rpm/find-debuginfo.sh --strict-build-id ...
    [kernel] *** ERROR: No build ID note found in /builddir ... vmlinuz

Unlike x86 bzImage, the power zImage is an elf module.  It is renamed
vmlinuz by rpmbuild via kernel.spec.  The fix is to specify the ld
command argument --build-id to inject the build-id into the target
image.  However, this is only done if the source image includes the
build-id parameter in the elf .notes section.

Change-Id: I57b5336dd6bef2e62e66e0444d8f640161730df3
Signed-off-by: Luke Browning <918375@nutanix.com>

Conflicts:
        arch/powerpc/boot/wrapper