ENG-211165 - PCI device series 8 12G SAS/PCIe3 secondary bus reset
The AACRAID device pretends that it has the FLR support and
does not work when host initiates the FLR. This change adds
the quirk for the device to use the secondary bus reset.

This quirk is different from other AHV SBR quirks, as there is
evidence FLR is completely broken for this device.

Testing:
 - The change is being tested on AHV with the device attached.
 - Booted with the new kernel image for a sanity check.

Change-Id: I2bb80f22378ffbd56f1472eedf12fe8843ffc8dd