Change CONFIG_HZ to 100 for powerpc
Power has more qemu kvm cpu threads as the VM runs in SMT-4 mode.  These threads
are running over fewer cpus in the host as Power8 is run with SMT disabled, so
there is a lot of context switching amongst these threads to schedule idle processors
in the VM.  The change to 1000 HZ is a recent one.  It was made to be more like
x86, but in this case it is not desirable with twice the threads and half the
processors, it slows down power as it introduces a lot of system overhead in the
host kernel.

Change-Id: Iaead3a06fdda292a1ec5a893f16b7ea0f4d5cb9e
Signed-off-by: Luke Browning [IBM] <918375@nutanix.com>