ENG-266899 - Update kernel config to ensure TSX is still enabled
We need the CPUD to still advertise HLE + RTM support so that VM's with
those features can live migrate to the host. The kernels "auto" and
"off" migitations set a MSR bit which results in HLE + RTM not being
advertised in CPUID and thus will cause to libvirt to not detect those
features.

The kernel use's the "verw" mitigation if TSX is left enabled and thus
the host is still secure.
https://software.intel.com/security-software-guidance/insights/deep-dive-intel-transactional-synchronization-extensions-intel-tsx-asynchronous-abort

Change-Id: Ide5a751bf4998089563d4333b19f8cb9656b19d9