ENG-282079: Add quirks for SMC/LSI HBA 3808/3816
On platforms with the SMC/LSI HBA 3808/3816, intermittent CVM boot
failures have been seen because the kernel fails to properly reset
the device before binding to the vfio module. The reset fails
because the device does not support Secondary Bus Resets (SBRs).
These failures are resolved by adding a PCI quirk to reset the
device using a Function Level Reset (FLR).

In preparation for the G8 platform, which will ship with the SMC/LSI
HBA 3808/3816, this change adds the aforementioned quirk to the AHV
kernel.

Change-Id: Id98083be6f8202d4bdcb6b0dc85d175fe951f380
Signed-off-by: Raphael Norwitz <raphael.norwitz@nutanix.com>
Signed-off-by: Alay Shah <alay.shah@nutanix.com>