sanity check avail_idx at VHOST_GET_VRING_BASE time
When userspace asks for the VRING_BASE (which in vhost is ->last_avail_idx),
sanity check its value against the guest ring's value, similar to the check in
vhost_get_vq_desc().

The main reason to add this is to help root cause the issue with the virtio
Windows drivers described in ENG-346611; as this is somewhat of a debugging
special for this as-yet-undiagnosed issue, it's not really suitable for
upstreaming.

Signed-off-by: John Levon <john.levon@nutanix.com>
Change-Id: I5082fe98c432a1429cb9409727e2e0eb703c6b14