ENG-380291: net: usbnet: allow overriding of default USB interface naming
When the predictable device naming scheme for NICs is not in use, it is
common for there to be udev rules to rename interfaces to names with
prefix "eth".

Since the timing at which USB NICs are discovered is unpredictable, it
can be interfere with udev's attempt to rename another interface to
"eth0" if a freshly discovered USB interface is initially given the name
"eth0".

Hence it is useful to be able to override the default name. A new usbnet
module parameter allows this to be configured.

Signed-off-by: Jonathan Davies <jonathan.davies@nutanix.com>
Suggested-by: Prashanth Sreenivasa <prashanth.sreenivasa@nutanix.com>

This patch was not accepted upstream, with the consensus being that
userspace needs to deal with the possibility of overlap between the
names initially assigned by the kernel and the intended names.

In other words, it's a result of our decision to use el7's systemd-udevd
(which assumes the consistent device naming scheme is being used) and
our use of the names Foundation selects, with prefix "eth" (which is
opposed to the consistent device naming scheme).

We cannot easily change systemd-udevd to bring back el6's backoff &
retry logic, and we cannot easily move away from NIC names having prefix
"eth". So it seems best to go with this patch as an interim solution
locally until we can move away from using NIC names with prefix "eth".
At that point, this patch can be dropped from our patch queue.

(cherry picked from commit 74bc5c48e0ff167667de67659585f0920d58877e)

Note: this commit was originally merged to linux-5.4.y-nutanix as commit
74bc5c48e. A subsequent force-push reverted the branch back to that
commit's parent. Hence that commit was lost, so it is being re-applied
here.

Change-Id: I004943898b337b574e63089399a69c6431cb730b