ENG-388414: Disable global reclamation from anonymous memory
Disable swapping of anonymous memory from memory cgroups with zero
swappiness *under global memory pressure*. This is achieved by assigning
SCAN_FILE to scan_balance inside get_scan_count() that stops considering
anonymous active and inactive LRU lists in the switch..case. Also, leave
a configuration knob in sysctl to revert to the previous default
behaviour.

This patch is intended to prevent swapping of the system daemons for
memory overcommit in the hosts with the swap attached via iSCSI.
Otherwise, the hosts are vulnerable to 1) a chicken and egg situation
when a swapped anonymous page is required to page it in, and
2) termination with SIGBUS of the paged out processes in case of
HA scenarios when the host swap becomes unavailable, e.g. fencing or
a network outage.

Previously, we reset the swappiness for all userspace processes by
resetting it with sysctl via kernel command-line option, which is later
inherited by all memory cgroups, and set it to non-zero for the OC VMs.
However, under global memory pressure, the Linux kernel can still
reclaim from the "protected" processes.

Alternatively, we could prevent swapping with mlockall with MCL_CURRENT
and MCL_FUTURE in the userspace, but this is troublesome because we
would need to patch many software components, including those taken
from the CentOS mirror, e.g. systemd.

The suggested patch does not affect the non-memory overcommitted hosts.
They should invoke the OOM killer as usual when the host can't cope
with memory pressure.

Change-Id: Icbef2025402cabe085f44f6e27ae1c013c5173b4

Conflicts:
	kernel/sysctl.c
	mm/vmscan.c