ENG-394311 kernel: enable CONFIG_BPF_JIT_ALWAYS_ON=y to nuke ___bpf_prog_run hook
In a previous change, we turned on a bunch-o-BPF options (great); however,
these have a side effect of making ___bpf_prog_run execute on every single
network transaction, which you can see in a vhost thread flamegraph.

To nuke this, set CONFIG_BPF_JIT_ALWAYS_ON=y, which removes the entire
BPF interpreter from the kernel and permanently enables the JIT compiler.

This gets rid of the ___bpf_prog_run hook and saves the CPU time for
steady state use cases. This will be good for both CVM and UVMs. I tested
this out and it works as expected.

Fixes: kernel ___bpf_prog_run hook executes on every single network transaction even when no BPF progs are present

Change-Id: I8b47b9eefab69c7eab525ac44421af2735ceae16