ENG-445963: Ensure MSR access support is always available
Libvirt 6.0 uses the interface provided by msr.ko to check MSRs. If it
can't read them, it uses fallback logic that leads to the pschange-mc-no
flag listed as a CPU feature in "virsh capabilities". This causes issues
when acropolis tries to start a VM with that flag, because the host CPU
might actually not support this flag after all.

Since we now require this interface for correct functionality, build
it directly into the kernel binary, so it is always available. This
seems both more straightforward and more robust than other options, such
as automatic on-demand loading via udev rules.

Change-Id: If48da5360043a873dd7f517a563b873850fb7b6f