Makefile.perf: use install -p
This parameter preserves the mtime of files being installed rather than
using the time of installation.

This makefile is used as part of the "install-tools" target of the main
Makefile.

Without this, the scripts scripts/python/*.py installed into
/usr/libexec/perf-core/scripts/python will have non-deterministic
timestamps, so when they are byte-compiled by rpmbuild, the .pyo and
.pyc files are non-deterministic.

Change-Id: I43b8ab97795e80a359fa842144c0dcadf4a8ff56