ENG-452882: Fix madvise(MADV_COLD) for shared memory
It's a workaround to improve the live migration of MemOC VMs. It's not
upstreamed.

QEMU uses madvise(MADV_COLD) during the memory scan for the live
migration to deactivate zero-pages that appeared due to the memory
scan. It helps preserve the guest's working set, decrease host swap
use and migration time.

With kernel 5.4 madvise(MADV_COLD) used to work as expected: the newly
allocated anon/shmem pages started on the active LRU, madvise(MADV_COLD)
called deactivate_page() to move them to the inactive LRU, and hence more
probable reclamation target.

However, it has changed around kernel 5.9. With commit b518154e59a
("mm/vmscan: protect the workingset on anonymous LRU"), anon/shmem pages
start on *inactive* LRU. In this case, QEMU -> madvise(MADV_COLD) ->
deactivate_page() is a NOP, which's unexpected and undesirable.

This patch makes the deactivate_page() move inactive pages at the end of
the inactive LRU and thus making them "cooler" than others, which are
genuinely the guest's working set in the inactive LRU.

Using the test program mentioned in the linux-mm thread below, no longer
observe swapping with newer kernels:

https://lore.kernel.org/linux-mm/CAOUHufaBi5Js6HXCO3R5nOo-+NiWajmDjM4mpx6H55JWS3auGQ@mail.gmail.com/#t

A move advanced version of this program:

https://github.com/nutanix-engineering/ahv-devtools/blob/main/memoc/test-programs/shmem.c

A patch for lru_deactivate_file_fn() suggested in the mailing list is
dropped because it seems irrelevant due to the memory type (anon/shmem,
not file-backed) and doesn't change the test case behaviour.

Change-Id: Ifd8c78b32e2bbc68796d46f0bf2f6d88caf81bb4