tools/kvm/kvm_stat: clamp timestamps in man page to SOURCE_DATE_EPOCH
kvm_stat.1 is generated using a2x which is a wrapper around asciidoc and
xsltproc.

By default, it is generated a US-format date in two places in the first
few lines of the file.

Sadly a2x doesn't respect SOURCE_DATE_EPOCH so we need to explicitly
override the date attribute on the command-line in order to build
kvm_stat.1 reproducibly.

(Although using ISO-8601 date would be preferable to US-format, the
hyphens get escaped which looks odd, so maintain the original format.)

Change-Id: I647a519e25285fe47ff0cb7b42279720e7b3e030