ENG-452323: Add ability to use the legacy start_pfn VFIO migration API
Add a QEMU_42_COMPAT_MIGRATION option to the kernel headers for vfio.h so that
we can easily force the Nvidia driver to use the old legacy start_pfn live
migration interface. This is because the new upstreamed version is not supported
by QEMU 4.2 which uses an outdated snapshot of the vfio.h definitions.

The build system of the Nvidia drivers try building a test program against this
header to determine whether or not the start_pfn member is part of the struct
and determines what migration API to compile in using the result of this build.

The QEMU headers are an old copy in that submodule of this header, but
since those headers are not installed or linked against by anything, the
approach of overriding the kernel header for this part of the API is simpler
and safer. These methods are not used by the Linux kernel at all.

Furthermore, there is a new notifier API used by the Nvidia drivers not present
in the QEMU copy of the headers but this API is not optional for the latest
drivers so we must link against the latest VFIO kernel UAPI header.

Change-Id: If23f97e6071d1277254c3d917657c6e9535e5de1