KVM: optionally commit write on ioeventfd write
This patch is a slightly different take on the ioregionfd mechanism
previously described here:
https://lore.kernel.org/all/88ca79d2e378dcbfb3988b562ad2c16c4f929ac7.camel@gmail.com/

The goal of this new mechanism, which we tentatively call shadow
ioeventfd in lack of a better name, is to avoid an exit to the VMM when
a particular GPA is written to. Without this mechanism, an exit to the
VMM is necessary so that it communicates the event, along with the value
written, to whoever is interested in this GPA.

Our motivation behind this to speed up doorbell writes on NVMe
controllers emulated outside of the VMM. Currently, a doorbell write to
an NVMe SQ tail doorbell requires returning from ioctl(KVM_RUN) and the
VMM communicating the event, along with the doorbell value, to the NVMe
controller emulation task.  With the shadow ioeventfd, the NVMe
emulation task is directly notified of the doorbell write and can find
the doorbell value in a known location, without the interference of the
VMM.

RFC posted to KVM mailing list: https://marc.info/?l=kvm&m=166500585028727&w=2

Set KVM_CAP_SHADOW_IOEVENTFD to a high numer (990) per ENG-532186 to
avoid clashes in the foreseeable future.

Signed-off-by: Thanos Makatos <thanos@nutanix.com>